Tested auth, updated models to fit with database

+++ src/main/java/com/project/controller/FileController.java
@@ -20,22 +20,22 @@ public class FileController {
     }
 
     @GetMapping("/{id}")
-    public FileEntity getFile(@PathVariable Long id) {
+    public FileEntity getFile(@PathVariable Integer id) {
         return fileService.getFileById(id);
     }
 
     @GetMapping("/task/{taskId}")
-    public List<FileEntity> getFilesByTask(@PathVariable Long taskId) {
+    public List<FileEntity> getFilesByTask(@PathVariable Integer taskId) {
         return fileService.getFilesByTask(taskId);
     }
 
     @GetMapping("/user/{userId}")
-    public List<FileEntity> getFilesByUser(@PathVariable Long userId) {
+    public List<FileEntity> getFilesByUser(@PathVariable Integer userId) {
         return fileService.getFilesByUser(userId);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteFile(@PathVariable Long id) {
+    public void deleteFile(@PathVariable Integer id) {
         fileService.deleteFile(id);
     }
 }
