+++ src/main/java/com/project/interfaces/IProjectMemberService.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface IProjectMemberService {
     ProjectMember addMember(ProjectMember member);
-    List<ProjectMember> getMembersByProject(Long projectId);
-    List<ProjectMember> getProjectsByUser(Long userId);
-    void removeMember(Long id);
-    ProjectMember updateRole(Long id, String role);
+    List<ProjectMember> getMembersByProject(Integer projectId);
+    List<ProjectMember> getProjectsByUser(Integer userId);
+    void removeMember(Integer id);
+    ProjectMember updateRole(Integer id, String role);
 
 }
