+++ src/main/java/com/project/repository/FileRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface FileRepository extends JpaRepository<FileEntity, Long> {
+public interface FileRepository extends JpaRepository<FileEntity, Integer> {
 
-    List<FileEntity> findByTaskId(Long taskId);
+    List<FileEntity> findByTaskId(Integer taskId);
 
-    List<FileEntity> findByUploadedById(Long userId);
+    List<FileEntity> findByUploadedById(Integer userId);
 }
